modified

+++ app/Http/Controllers/Backend/ProgramsController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Actions\SlugChecker;
 use App\Models\ProgramRegistration;
+use App\Models\ProgramPaymentLogs;
 use App\Models\Programs;
 use App\Facades\EmailTemplate;
+use App\Helpers\ProjectHelpers;
 use App\Http\Requests\SlugCheckRequest;
 use App\Mail\BasicMail;
 use App\Mail\RegistrationReply;
-use App\Helpers\ProjectHelpers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Str;
@@ -31,10 +32,10 @@ class ProgramsController extends Controller
             'program_content' => 'required|string',
             'venue' => 'nullable|string',
             'venue_location' => 'nullable|string',
-            'venue_phone' => 'nullable|string',
             'time' => 'required|string',
             'image' => 'nullable|string',
             'date' => 'required|string',
+            'cost' => 'required|string',
             'available_registrations' => 'required|string',
             'slug' => 'nullable|string'
         ]);
@@ -49,6 +50,7 @@ class ProgramsController extends Controller
             'status' => $request->status,
             'date' => $request->date,
             'time' => $request->time,
+            'cost' => $request->cost,
             'available_registrations' => $request->available_registrations,
             'image' => $request->image,
             'venue' => $request->venue,
@@ -57,7 +59,7 @@ class ProgramsController extends Controller
             'meta_description' => $request->meta_description,
         ]);
 
-        return redirect()->back()->with(['msg' => __('New Event Created Success...'),'type'=>'success']);
+        return redirect()->back()->with(['msg' => __('New Program Created Success...'),'type'=>'success']);
     }
 
     public function all_programs(){
@@ -72,7 +74,7 @@ class ProgramsController extends Controller
 
     public function delete_program(Request $request,$id){
         Programs::find($id)->delete();
-        return redirect()->back()->with(['msg' => __('Event Delete Success...'),'type'=>'danger']);
+        return redirect()->back()->with(['msg' => __('Program Delete Success...'),'type'=>'danger']);
     }
 
     public function update_program(Request $request){
@@ -84,6 +86,7 @@ class ProgramsController extends Controller
             'time' => 'required|string',
             'image' => 'nullable|string',
             'date' => 'required|string',
+            'cost' => 'required|string',
             'available_registrations' => 'required|string',
             'slug' => 'nullable|string'
         ]);
@@ -97,6 +100,7 @@ class ProgramsController extends Controller
             'status' => $request->status,
             'date' => $request->date,
             'time' => $request->time,
+            'cost' => $request->cost,
             'available_registrations' => $request->available_registrations,
             'image' => $request->image,
             'venue' => $request->venue,
@@ -105,7 +109,7 @@ class ProgramsController extends Controller
             'meta_description' => $request->meta_description,
         ]);
 
-        return redirect()->back()->with(['msg' => __('Event Update Success...'),'type'=>'success']);
+        return redirect()->back()->with(['msg' => __('Program Update Success...'),'type'=>'success']);
     }
 
     public function clone_program(Request $request){
@@ -117,12 +121,13 @@ class ProgramsController extends Controller
             'status' => 'draft',
             'date' => $program_details->date,
             'time' => $program_details->time,
+            'cost' => $program_details->cost,
             'available_registrations' => $program_details->available_registrations,
             'image' => $program_details->image,
             'venue' => $program_details->venue,
             'venue_location' => $program_details->venue_location,
         ]);
-        return redirect()->back()->with(['msg' => __('Events Clone Success...'),'type' => 'success']);
+        return redirect()->back()->with(['msg' => __('Programs Clone Success...'),'type' => 'success']);
     }
 
     public function program_registration(){
@@ -157,6 +162,10 @@ class ProgramsController extends Controller
 
     public function delete_program_registration_logs(Request $request,$id){
         $registration_details = ProgramRegistration::find($id);
+        $program_payment_logs = ProgramPaymentLogs::where('registration_id',$registration_details->id)->first();
+        if (!empty($program_payment_logs)){
+            return redirect()->back()->with(['msg' => __('Your Can not delete this registration, it already associated with a program payment log.'),'type' => 'danger']);
+        }
         $registration_details->delete();
         return redirect()->back()->with(['msg' => __('Programs Registration Lob Deleted...'),'type' => 'danger']);
     }
@@ -166,14 +175,14 @@ class ProgramsController extends Controller
         
          //todo: write code to increase  ticket number if status == cancel
         if($request->registration_status == 'canceled'){
-            //update program available tickets
+            //update program available registrations
             $registration_details = ProgramRegistration::where('id',$request->registration_id)->first();
             $program_details = Programs::findOrFail($registration_details->program_id);
             $program_details->available_registrations = (int) $program_details->available_registrations + $registration_details->quantity;
             $program_details->save();
         }
         
-        return redirect()->back()->with(['msg' => __('Program Registration Status Updated...'),'type' => 'success']);
+        return redirect()->back()->with(['msg' => __('Programs Registration Status Updated...'),'type' => 'success']);
     }
 
     public function send_mail_program_registration_logs(Request $request){
@@ -193,7 +202,45 @@ class ProgramsController extends Controller
         }catch (\Exception $e){
             return redirect()->back()->with(ProjectHelpers::item_delete($e->getMessage()));
         }
-        return redirect()->back()->with(['msg' => __('Attendance Reply Mail Send Success...'),'type' => 'success']);
+        return redirect()->back()->with(['msg' => __('Registration Reply Mail Send Success...'),'type' => 'success']);
+    }
+
+    public function program_payment_logs(){
+        $paymeng_logs = ProgramPaymentLogs::all();
+        return view('backend.programs.program-payment-logs-all')->with(['payment_logs' => $paymeng_logs]);
+    }
+    public function delete_program_payment_logs(Request $request,$id){
+        ProgramPaymentLogs::find($id)->delete();
+        return redirect()->back()->with(['msg' => __('Program Payment Log Delete Success...'),'type' => 'danger']);
+    }
+
+    public function approve_program_payment(Request $request,$id){
+
+        $payment_logs = ProgramPaymentLogs::find($id);
+        $payment_logs->status = 'complete';
+        $payment_logs->save();
+
+        $program_registration = ProgramRegistration::find($payment_logs->registration_id);
+        $program_registration->payment_status = 'complete';
+        $program_registration->status = 'complete';
+        $program_registration->save();
+
+        $program_details = Programs::find($program_registration->program_id);
+        $program_details->available_registrations -= $program_registration->quantity;
+        $program_details->save();
+
+        //update database
+        $program_payment_logs = ProgramPaymentLogs::find($id);
+        $program_registration = ProgramRegistration::find($program_payment_logs->registration_id);
+
+        $order_mail = get_static_option('program_registration_receiver_mail') ?: get_static_option('site_global_email');
+        try {
+            Mail::to($order_mail)->send(new BasicMail(EmailTemplate::programRegistrationPaymentAcceptMail($program_registration)));
+        }catch (\Exception $e){
+            return redirect()->back()->with(['msg' => __('Manual Payment Accept Success, mail send failed').' '.$e->getMessage(),'type' => 'success']);
+        }
+
+        return redirect()->back()->with(['msg' => __('Manual Payment Accept Success'),'type' => 'success']);
     }
 
     public function bulk_action(Request $request){
@@ -206,9 +253,43 @@ class ProgramsController extends Controller
         return response()->json(['status' => 'ok']);
     }
 
+    public function payment_logs_bulk_action(Request $request){
+        ProgramPaymentLogs::whereIn('id',$request->ids)->delete();
+        return response()->json(['status' => 'ok']);
+    }
+    
+    public function payment_report(Request  $request){
+        $order_data = '';
+        $query = ProgramPaymentLogs::query();
+        if (!empty($request->start_date)){
+            $query->whereDate('created_at','>=',$request->start_date);
+        }
+        if (!empty($request->end_date)){
+            $query->whereDate('created_at','<=',$request->end_date);
+        }
+        if (!empty($request->payment_status)){
+            $query->where(['status' => $request->payment_status ]);
+        }
+        $error_msg = __('select start & end date to generate program payment report');
+        if (!empty($request->start_date) && !empty($request->end_date)){
+            $query->orderBy('id','DESC');
+            $order_data =  $query->paginate($request->items);
+            $error_msg = '';
+        }
+
+        return view('backend.programs.payment-report')->with([
+            'order_data' => $order_data,
+            'start_date' => $request->start_date,
+            'end_date' => $request->end_date,
+            'items' => $request->items,
+            'payment_status' => $request->payment_status,
+            'error_msg' => $error_msg
+        ]);
+    }
+
     public function registration_report(Request  $request){
         $order_data = '';
-        $programs = Programs::where(['status' => 'publish'])->get();
+        $programs = Programs::get();
         $query = ProgramRegistration::query();
         if (!empty($request->start_date)){
             $query->whereDate('created_at','>=',$request->start_date);
@@ -237,18 +318,23 @@ class ProgramsController extends Controller
         ]);
     }
 
-    public function program_attedance_reminder(Request $request){
+    public function program_registration_reminder(Request $request){
         //send order reminder mail
-        $registration_details = ProgramRegistration::find($request->id);
-        $send_to = ProgramRegistration::where(['registration_id' => $request->id ])->first();
+        $order_details = ProgramRegistration::find($request->id);
+        $payment_log = ProgramPaymentLogs::where(['registration_id' => $request->id])->first();
         try {
-            Mail::to($send_to->email)->send(new BasicMail(EmailTemplate::programBookingReminderMail($registration_details)));
+            Mail::to($payment_log->email)->send(new BasicMail(EmailTemplate::programRegistrationReminderMail($order_details)));
         }catch (\Exception $e){
             return back()->with(ProjectHelpers::item_delete($e->getMessage()));
         }
 
         return redirect()->back()->with(['msg' => __('Reminder Mail Send Success'),'type' => 'success']);
     }
+
+    public function settings(){
+        return view('backend.programs.settings');
+    }
+
     public function slug_check(SlugCheckRequest $request){
 
         $user_given_slug = $request->slug;
